[core] Power to the primitive value
- new value::primitive module
- move dicom_value! macro to lib root
- new error type ConvertValueError
   - add variant to existing error types
- narrow errors of result types from value::deserialize
- improve documentation and behavior of dicom_value!
- improve documentation of PrimitiveValue
- impl From<Vec<u8>> for PrimitiveValue
- impl From<&[u8]> for PrimitiveValue
- impl Display for PrimitiveValue
- add PrimitiveValue::to_str
- add PrimitiveValue::to_bytes
   - required safe-transmute dependency for
   the binary slice to bytes conversion
- add PrimitiveValue::to_date
