+++ core/Cargo.toml
@@ -18,3 +18,4 @@ chrono = "0.4.6"
 itertools = "0.9.0"
 quick-error = "1.2.3"
 smallvec = "1.0.0"
+safe-transmute = "0.11.0"
